Transpile and seperation of script
Added babel and it's need dependencies for transpiling src scripts to the lib ES2015 script files. //0: Keep needed to be added in places to keep the map intact from the forEach plugin.

Also seperated the script for whatcom data and the generic to two seperate scrip files. Just will need to repoint the path in index if depending on what script to use